+++ src/main/java/com/vasiliy/project/service/PredictionProductService.java
@@ -10,7 +10,7 @@ public interface PredictionProductService {
 
   PredictionDataDTO getPredictionDTO(Long productId, Integer numberOfLastWeeks);
 
-  Integer getNextWeekPrediction(List<Integer> weekOutflowValues);
+  Double getNextWeekPrediction(List<Integer> weekOutflowValues);
 
-  Integer getNextMonthPrediction(List<Integer> monthOutflowValues);
+  Double getNextMonthPrediction(List<Integer> monthOutflowValues);
 }
